feat: 예약 API 연동 및 실패 시 화면 초기화

+++ app/src/main/java/com/example/popmate/model/data/remote/reservation/CurrentReservationResponse.kt
@@ -6,7 +6,7 @@ data class CurrentReservationResponse(
     val popupStoreDescription: String,
     val popupStoreOpenTime: String,
     val popupStoreTitle: String,
-    val reservationId: Int,
+    val reservationId: Long,
     val startTime: String,
     val status: String
 )
